Feat : 카테고리별 비디오 api 완료

+++ backend/src/main/java/com/ssafy/cnnect/video/repository/VideoRepository.java
@@ -2,6 +2,8 @@ package com.ssafy.cnnect.video.repository;
 
 import com.ssafy.cnnect.video.dto.VideoListResponseDto;
 import com.ssafy.cnnect.video.entity.Video;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 
@@ -10,8 +12,10 @@ import java.util.List;
 public interface VideoRepository extends MongoRepository<Video, String> {
     @Query("{}")
     List<Video> findAllVideo();
+
     @Query("{category_id : ?0}")
-    List<Video> findByCategoryId(int categoryId);
+    List<Video> findByCategoryId(Long categoryId);
+
     @Query("{video_id : ?0}")
     Video findByVideoId(String videoId);
 }
